@@ -1,7 +1,7 @@
 package com.ramotion.cardslider;
 
-import android.support.annotation.NonNull;
-import android.support.v4.view.ViewCompat;
+import androidx.annotation.NonNull;
+import androidx.core.view.ViewCompat;
 import android.view.View;
 
 /**
