@@ -4,14 +4,15 @@ import android.animation.ObjectAnimator;
 import android.graphics.Bitmap;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.CardView;
 import android.transition.Transition;
 import android.util.DisplayMetrics;
 import android.view.View;
 import android.widget.ImageView;
 
 import com.ramotion.cardslider.examples.simple.utils.DecodeBitmapTask;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.cardview.widget.CardView;
 
 public class DetailsActivity extends AppCompatActivity implements DecodeBitmapTask.Listener {
 
