@@ -1,12 +1,13 @@
 package com.ramotion.cardslider.examples.simple.cards;
 
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.ramotion.cardslider.examples.simple.R;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 public class SliderAdapter extends RecyclerView.Adapter<SliderCard> {
 
