@@ -1,14 +1,15 @@
 package com.ramotion.cardslider.examples.simple.cards;
 
 import android.graphics.Bitmap;
-import android.support.annotation.DrawableRes;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.view.ViewTreeObserver;
 import android.widget.ImageView;
 
 import com.ramotion.cardslider.examples.simple.R;
 import com.ramotion.cardslider.examples.simple.utils.DecodeBitmapTask;
+
+import androidx.annotation.DrawableRes;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class SliderCard extends RecyclerView.ViewHolder implements DecodeBitmapTask.Listener {
 
