@@ -11,13 +11,14 @@ import android.graphics.Rect;
 import android.graphics.RectF;
 import android.os.AsyncTask;
 import android.os.Build;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.NonNull;
 
 import com.ramotion.cardslider.examples.simple.R;
 
 import java.lang.ref.Reference;
 import java.lang.ref.WeakReference;
+
+import androidx.annotation.DrawableRes;
+import androidx.annotation.NonNull;
 
 
 public class DecodeBitmapTask extends AsyncTask<Void, Void, Bitmap> {
